fix: make labels float instead of long tensor

+++ dataset.py
@@ -10,7 +10,7 @@ class BlogDataset(Dataset):
 
     def __getitem__(self, index):
         row = self.data.iloc[index]
-        return row["body"], torch.tensor(row[2:])
+        return row["body"], torch.tensor(row[2:]).to(float)
 
     def __len__(self):
         return len(self.data)
